% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{tablerAvatarList}
\alias{tablerAvatarList}
\title{Create a Boostrap 4 avatar list}
\usage{
tablerAvatarList(..., stacked = FALSE)
}
\arguments{
\item{...}{Slot for \link{tablerAvatar}.}

\item{stacked}{Whether to stack avatars. FALSE by default.}
}
\description{
Build a tabler avatar list
}
\examples{
if(interactive()){
 library(shiny)
 library(tablerDash)

 shiny::shinyApp(
   ui = tablerDashPage(
    navbar = NULL,
    footer = NULL,
    title = "test",
    body = tablerDashBody(
     tablerAvatarList(
      stacked = TRUE,
      tablerAvatar(
       name = "DG",
       size = "xxl"
      ),
      tablerAvatar(
       name = "DG",
       color = "orange"
      ),
      tablerAvatar(
       name = "DG",
       status = "warning"
      ),
      tablerAvatar(url = "https://image.flaticon.com/icons/svg/145/145852.svg")
     )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
