% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBody.R
\name{tablerDashBody}
\alias{tablerDashBody}
\title{Create a Boostrap 4 dashboard body}
\usage{
tablerDashBody(...)
}
\arguments{
\item{...}{Body content, slot for \link{tablerTabItems}.}
}
\description{
Build a tabler dashboard body
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
