% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{tablerDashGallery}
\alias{tablerDashGallery}
\title{Launch the tablerDash Gallery}
\usage{
tablerDashGallery()
}
\description{
A gallery of all components available in tablerDash.
}
\examples{

if (interactive()) {

 tablerDashGallery()

}
}
