% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{tablerDashNav}
\alias{tablerDashNav}
\title{Create a Boostrap 4 dashboard navbar}
\usage{
tablerDashNav(id, ..., src = NULL, navMenu = NULL)
}
\arguments{
\item{id}{Navbar id.}

\item{...}{Navbar content.}

\item{src}{Brand image url or path.}

\item{navMenu}{Slot for \link{tablerNavMenu}}
}
\description{
Build a tabler dashboard page
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
