% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardPage.R
\name{tablerDashPage}
\alias{tablerDashPage}
\title{Create a Boostrap 4 dashboard page}
\usage{
tablerDashPage(navbar = NULL, body = NULL, footer = NULL,
  title = NULL, enable_preloader = FALSE, loading_duration = 2)
}
\arguments{
\item{navbar}{Slot for \link{tablerDashNav}.}

\item{body}{Slot for \link{tablerDashBody}.}

\item{footer}{Slot for \link{tablerDashFooter}.}

\item{title}{App title.}

\item{enable_preloader}{Whether to enable a page loader. FALSE by default.}

\item{loading_duration}{Loader duration in seconds. 2s by default.}
}
\description{
Build a tabler dashboard page
}
\examples{
if(interactive()){
 library(shiny)
 library(tablerDash)

 shiny::shinyApp(
   ui = tablerDashPage(
    navbar = tablerDashNav(),
    footer = tablerDashFooter(),
    title = "test",
    body = tablerDashBody()
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
