% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{tablerDropdown}
\alias{tablerDropdown}
\title{Create a Boostrap 4 dashboard dropdown container}
\usage{
tablerDropdown(..., icon = "bell", collapsed = TRUE)
}
\arguments{
\item{...}{Slot for \link{tablerDropdownItem}.}

\item{icon}{Dropdown icon.}

\item{collapsed}{Whether to collapse the dropdown. TRUE by default.}
}
\description{
Build a tabler dashboard dropdown container
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
