% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{tablerDropdownItem}
\alias{tablerDropdownItem}
\title{Create a Boostrap 4 dashboard dropdown container}
\usage{
tablerDropdownItem(..., title = NULL, href = NULL, url = NULL,
  status = NULL, date = NULL)
}
\arguments{
\item{...}{Item content.}

\item{title}{Item title.}

\item{href}{External link.}

\item{url}{Author image.}

\item{status}{Item status. Only if url is not NULL.}

\item{date}{Item date.}
}
\description{
Build a tabler dashboard dropdown container
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
