% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{tablerIcon}
\alias{tablerIcon}
\title{Create a Boostrap 4 icon}
\usage{
tablerIcon(name, lib = c("feather", "font-awesome", "flag", "payment"))
}
\arguments{
\item{name}{Name of icon. See \url{https://preview.tabler.io/icons.html} for all valid icons.}

\item{lib}{Icon library ("feather", "font-awesome", "flag", "payment").}
}
\description{
Build a tabler icon
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
