% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{tablerList}
\alias{tablerList}
\title{Create a Boostrap 4 list container}
\usage{
tablerList(...)
}
\arguments{
\item{...}{Slot for \link{tablerListItem}.}
}
\description{
Build a tabler list container
}
\examples{
if(interactive()){
 library(shiny)
 library(tablerDash)

 shiny::shinyApp(
   ui = tablerDashPage(
    navbar = NULL,
    footer = NULL,
    title = "test",
    body = tablerDashBody(
      tablerCard(
       title = "tablerList",
       tablerList(
       tablerListItem(
        tablerTag(name = "Tag"),
        tablerTag(name = "Tag", href = "https://www.google.com"),
        tablerTag(name = "Tag", rounded = TRUE, color = "pink")
       ),
       tablerListItem(tablerStatus(color = "red")),
       tablerListItem(
        tablerAvatarList(
         stacked = TRUE,
         tablerAvatar(
           name = "DG",
           size = "xxl"
         ),
         tablerAvatar(
           name = "DG",
           color = "orange"
         ),
         tablerAvatar(
           name = "DG",
           status = "warning"
         ),
         tablerAvatar(url = "https://image.flaticon.com/icons/svg/145/145852.svg")
         )
        )
       )
      )
     )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
