% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{tablerListItem}
\alias{tablerListItem}
\title{Create a Boostrap 4 list item}
\usage{
tablerListItem(...)
}
\arguments{
\item{...}{Slot for any HTML element.}
}
\description{
Build a tabler list item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
