% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{tablerMediaCard}
\alias{tablerMediaCard}
\title{Create a Boostrap 4 media card}
\usage{
tablerMediaCard(title = NULL, date = NULL, href = NULL, src = NULL,
  avatarUrl = NULL, width = 4)
}
\arguments{
\item{title}{Title.}

\item{date}{Date.}

\item{href}{External link.}

\item{src}{Image url.}

\item{avatarUrl}{Avatar image.}

\item{width}{Card width. 4 by default.}

\item{...}{Any other elements.}
}
\description{
Build a tabler media card
}
\examples{
if(interactive()){
 library(shiny)
 library(tablerDash)

 shiny::shinyApp(
   ui = tablerDashPage(
    navbar = NULL,
    footer = NULL,
    title = "test",
    body = tablerDashBody(
     tablerMediaCard(
      title = "Media Card",
      date = "Today",
      href = "https://www.google.com",
      src = "https://preview.tabler.io/demo/photos/matt-barrett-339981-500.jpg",
      avatarUrl = "https://image.flaticon.com/icons/svg/145/145842.svg",
      width = 6
     )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
