% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{tablerNavMenu}
\alias{tablerNavMenu}
\title{Create a Boostrap 4 dashboard navbar menu}
\usage{
tablerNavMenu(...)
}
\arguments{
\item{...}{Slot for \link{tablerNavMenuItem}.}
}
\description{
Build a tabler dashboard main navbar menu
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
