% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardNavbar.R
\name{tablerNavMenuItem}
\alias{tablerNavMenuItem}
\title{Create a Boostrap 4 dashboard navbar menu item}
\usage{
tablerNavMenuItem(..., tabName = NULL, icon = NULL)
}
\arguments{
\item{...}{Item name.}

\item{tabName}{Should correspond exactly to the tabName given in \code{\link{tablerTabItem}}.}

\item{icon}{Item icon.}
}
\description{
Build a tabler dashboard navbar menu item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
