% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{tablerProgress}
\alias{tablerProgress}
\title{Create a Boostrap 4 progress bar}
\usage{
tablerProgress(value, status = NULL, size = NULL)
}
\arguments{
\item{value}{Progress value.}

\item{status}{Progress status. See \url{https://preview.tabler.io/docs/colors.html}.}

\item{size}{Progress bar size: NULL, "xs", "sm", "md".}
}
\description{
Build a tabler progress bar
}
\examples{
if(interactive()){
 library(shiny)
 library(tablerDash)

 shiny::shinyApp(
   ui = tablerDashPage(
    navbar = NULL,
    footer = NULL,
    title = "test",
    body = tablerDashBody(
      tablerProgress(value = 10, size = "xs"),
      tablerProgress(value = 90, status = "red", size = "sm")
     )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
