% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{tablerSocialLink}
\alias{tablerSocialLink}
\title{Create a Boostrap 4 social link}
\usage{
tablerSocialLink(name = NULL, href = NULL, icon)
}
\arguments{
\item{name}{Link tooltip name.}

\item{href}{External link.}

\item{icon}{Icon (font awesome).}
}
\description{
Build a tabler social link
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
