% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{tablerSocialLinks}
\alias{tablerSocialLinks}
\title{Create a Boostrap 4 social link list}
\usage{
tablerSocialLinks(...)
}
\arguments{
\item{...}{Slot for \link{tablerSocialLink}.}
}
\description{
Build a tabler social link list
}
\examples{
if(interactive()){
 library(shiny)
 library(tablerDash)

 shiny::shinyApp(
   ui = tablerDashPage(
    navbar = NULL,
    footer = NULL,
    title = "test",
    body = tablerDashBody(
     tablerSocialLinks(
       tablerSocialLink(
        name = "facebook",
        href = "https://www.facebook.com",
        icon = "facebook"
       ),
       tablerSocialLink(
        name = "twitter",
        href = "https://www.twitter.com",
        icon = "twitter"
       )
      )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
