% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{tablerStatCard}
\alias{tablerStatCard}
\title{Create a Boostrap 4  stat card}
\usage{
tablerStatCard(value, title, trend = NULL, width = 3)
}
\arguments{
\item{value}{Card value.}

\item{title}{Card title.}

\item{trend}{Percentage increase/decrease.}

\item{width}{Card width. 3 by default.}
}
\description{
Build a tabler stat card
}
\examples{
if(interactive()){
 library(shiny)
 library(tablerDash)

 shiny::shinyApp(
   ui = tablerDashPage(
    navbar = NULL,
    footer = NULL,
    title = "test",
    body = tablerDashBody(
     tablerStatCard(
      value = 43,
      title = "Followers",
      trend = -10
     )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
