% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{tablerStatus}
\alias{tablerStatus}
\title{Create a Boostrap 4 status}
\usage{
tablerStatus(color)
}
\arguments{
\item{color}{Status color. See \url{https://preview.tabler.io/docs/colors.html}.}
}
\description{
Build a tabler status
}
\examples{
if(interactive()){
 library(shiny)
 library(tablerDash)

 shiny::shinyApp(
   ui = tablerDashPage(
    navbar = NULL,
    footer = NULL,
    title = "test",
    body = tablerDashBody(
     tablerStatus(color = "lime")
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
