% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBody.R
\name{tablerTabItems}
\alias{tablerTabItems}
\title{A container for tab items}
\usage{
tablerTabItems(...)
}
\arguments{
\item{...}{Items to put in the container. Each item should be a
\code{\link{tablerTabItem}}.}
}
\description{
A container for tab items
}
