% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{tablerTableItem}
\alias{tablerTableItem}
\title{Create a Boostrap 4 table item}
\usage{
tablerTableItem(left = NULL, right = NULL)
}
\arguments{
\item{left}{Left elements.}

\item{right}{Right elements.}
}
\description{
Build a tabler table item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
