% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabset.R
\name{tablerTabsetItem}
\alias{tablerTabsetItem}
\title{Tabler tabset item}
\usage{
tablerTabsetItem(title, ..., value = title, icon = NULL,
  title_header = NULL)
}
\description{
Tabler tabset item
}
\details{
Creating tabler tabset item.
}
