% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabset.R
\name{tablerTabsetPanel}
\alias{tablerTabsetPanel}
\title{Tabler tabset  panel}
\usage{
tablerTabsetPanel(..., id = NULL, selected = NULL, type = c("tabs"),
  width = 6)
}
\description{
Tabler tabset  panel
}
\details{
Generating tabset from tab items.
}
