% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{tablerTagList}
\alias{tablerTagList}
\title{Create a Boostrap 4 tag list}
\usage{
tablerTagList(...)
}
\arguments{
\item{...}{Slot for \link{tablerTag}.}
}
\description{
Build a tabler tag list
}
\examples{
if(interactive()){
 library(shiny)
 library(tablerDash)

 shiny::shinyApp(
   ui = tablerDashPage(
    navbar = NULL,
    footer = NULL,
    title = "test",
    body = tablerDashBody(
     tablerTagList(
      lapply(X = 1:5, FUN = function(i) {
       tablerTag(name = i)
      })
     )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
