% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{tablerTimeline}
\alias{tablerTimeline}
\title{Create a Boostrap 4 timeline}
\usage{
tablerTimeline(...)
}
\arguments{
\item{...}{slot for \link{tablerTimelineItem}.}
}
\description{
Build a tabler timeline
}
\examples{
if(interactive()){
 library(shiny)
 library(tablerDash)

 shiny::shinyApp(
   ui = tablerDashPage(
    navbar = NULL,
    footer = NULL,
    title = "test",
    body = tablerDashBody(
     tablerTimeline(
      tablerTimelineItem(
       title = "Item 1",
       status = "green",
       date = "now"
      ),
      tablerTimelineItem(
       title = "Item 2",
       status = NULL,
       date = "yesterday",
       "Lorem ipsum dolor sit amet,
       consectetur adipisicing elit."
      )
     )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
