% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{tablerTimelineItem}
\alias{tablerTimelineItem}
\title{Create a Boostrap 4 timeline item}
\usage{
tablerTimelineItem(..., status = NULL, title, date)
}
\arguments{
\item{...}{Item content.}

\item{status}{Item status.}

\item{title}{Item title.}

\item{date}{Event date.}
}
\description{
Build a tabler timeline item
}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
