% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{tablerCard}
\alias{tablerCard}
\title{Create a Boostrap 4 card}
\usage{
tablerCard(..., title = NULL, options = NULL, footer = NULL,
  status = NULL, statusSide = c("top", "left"))
}
\arguments{
\item{...}{Body content}

\item{title}{Card title. If NULL, the header is not displayed.}

\item{options}{Card extra header elements.}

\item{footer}{Card footer. NULL by default. Not displayed if NULL.}

\item{status}{Card status. NULL by default. See \link{https://preview.tabler.io/docs/colors.html} for valid statuses.}

\item{statusSide}{Status side: "top" or "left".}
}
\description{
Build a tabler card
}
\examples{
if(interactive()){
 library(shiny)
 library(tablerDash)

 shiny::shinyApp(
   ui = tablerDashPage(
    navbar = NULL,
    footer = NULL,
    title = "test",
    body = tablerDashBody(
     tablerCard(
      title = "Card",
      "Inside the card",
      status = "success",
      statusSide = "left"
     )
    )
   ),
   server = function(input, output) {}
 )
}

}
\author{
David Granjon, \email{dgranjon@ymail.com}
}
